+++ src/components/DoctorDashboard.jsx
@@ -94,9 +94,14 @@ const DoctorDashboard = () => {
   };
 
   const handleLogout = () => {
+    const confirmation_message_to_logout = "You are about to logout.\nProceed?"
+    if(!confirm(confirmation_message_to_logout)){
+      return
+    }
     localStorage.setItem('access_token', null);  
     localStorage.setItem('user_id', null);
     localStorage.setItem('user_role', null);
+
     navigate("/");
   };
