+++ src/components/PatientDashboard.jsx
@@ -288,12 +288,17 @@ const PatientDashboard = () => {
       // setMessage(`Error: ${error.message}`);
     }
   };
-
+  
   const handleLogout = () => {
-    localStorage.setItem('access_token', null);  // or username if you're saving that
+    const confirmation_message_to_logout = "You are about to logout.\nProceed?"
+    if(!confirm(confirmation_message_to_logout)){
+      return
+    }
+    localStorage.setItem('access_token', null);  
     localStorage.setItem('user_id', null);
     localStorage.setItem('user_role', null);
-    navigate("/"); // Redirect to landing page
+
+    navigate("/");
   };
 
   /**
